add task to generate ViewHolder
