+++ linen-glue/build.gradle
@@ -6,5 +6,6 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile project(':wheat-ancient')
+    compile 'com.android.support:recyclerview-v7:23.1.0'
     compile 'com.android.support:appcompat-v7:23.1.0'
 }
