+++ linen-glue/src/main/java/x7c1/linen/glue/res/layout/SourceRow.java
@@ -1,0 +1,27 @@
+
+/**
+ * This file is automatically generated by wheat-build.
+ * Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+ */
+
+package x7c1.linen.glue.res.layout;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+import android.widget.TextView;
+
+public class SourceRow extends RecyclerView.ViewHolder {
+
+    public final TextView title;
+    public final TextView description;
+
+    public SourceRow(
+        View itemView,
+        TextView title,
+        TextView description
+    ){
+        super(itemView);
+        this.title = title;
+        this.description = description;
+    }
+}
